Files needed for the optimization
However, there is some problem with matrix multiplication. Strangely, the problem is in the part we already did for the last project.
Rundeopt is the main file, the only things I changed are the number of parameters and maybe the strategy used. It then calls deopt which I didn't change except for the name of the objective function, deopt calls optionpriceopt which I write, optionpriceopt calls c_hat which is our estimation function from the last project but with different output.